+++ tests/snapshot_tests/test_snapshots.py
@@ -1236,7 +1236,7 @@ def test_example_pride(snap_compare):
     """Test the pride example."""
     assert snap_compare(EXAMPLES_DIR / "pride.py")
 
-    
+
 def test_button_with_console_markup(snap_compare):
     """Regression test for https://github.com/Textualize/textual/issues/4328"""
     assert snap_compare(SNAPSHOT_APPS_DIR / "button_markup.py")
@@ -1249,3 +1249,7 @@ def test_width_100(snap_compare):
 
 def test_button_with_multiline_label(snap_compare):
     assert snap_compare(SNAPSHOT_APPS_DIR / "button_multiline_label.py")
+
+
+def test_margin_multiple(snap_compare):
+    assert snap_compare(SNAPSHOT_APPS_DIR / "margin_multiple.py")