+++ CHANGELOG.md
@@ -11,6 +11,10 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 - Fixed `TextArea` to end mouse selection only if currently selecting https://github.com/Textualize/textual/pull/4436
 
+### Changed
+
+- Added argument to signal callbacks https://github.com/Textualize/textual/pull/4438
+
 ## [0.57.1] - 2024-04-20
 
 ### Fixed