+++ CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 - Fixed `TextArea` to end mouse selection only if currently selecting https://github.com/Textualize/textual/pull/4436
 - Fixed issue with scroll_to_widget
+- Fixed issue with margins https://github.com/Textualize/textual/pull/4441
 
 ### Changed